refactor: rewrite download status handling
resolves: #174

+++ app/src/main/java/ir/jaamebaade/jaamebaade_client/view/PoemScreen.kt
@@ -36,7 +36,7 @@ fun PoemScreen(
     navController: NavController,
     poemId: Int,
     poetId: Int,
-    focusedVerseId: Int?,
+    focusedVerseId: Long?,
     modifier: Modifier,
     audioViewModel: AudioViewModel = hiltViewModel()
 ) {
